Исправить обобщенные тесты TaskManagerTest Добавить логирование NumberFormatException
